requirements.txt
streamlit
openai>=1.0
python-dotenv
fpdf